[kuduraft] [proxy] explicitly clear proxy uuid from the request object
Summary:
The recipient of a consensus request message determines if the
request is a 'proxy' request by checking the 'proxy_dest_uuid' field of
the request. When this is set, then the message is a proxy request.
Based on this fact, this message will be rejected if
'proxy_dest_uuid' is not the same as the recipient's uuid. This is
essentially a safeguard against misrouted proxy requests.

The request object is cached for each peer and relevent fields are
rebuilt whent the leader has to send a request. When proxy gets disabled,
the 'proxy_dest_uuid' field was not cleared. The message is routed
directly to the destination, but this message has the 'proxy_dest_uuid'
set to the old 'proxy_peer'. This triggers the check mentioned above and
the recipient rejects this message and will never be able to recives any
messages from the leader. This PR fixes the problem by explicitly
clearing this field when building the request on the leader.

Test Plan: local tp2 checkouts and trigger this scenario in fbcode

Reviewers: arahut, mpercy

Subscribers:

Tasks:

Tags: